src/test/java/com/leetcode/arrays/medium/TheEarliestMomentWhenEveryoneBecomeFriends_1101Test.java
package com.leetcode.arrays.medium;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class TheEarliestMomentWhenEveryoneBecomeFriends_1101Test {

  private final TheEarliestMomentWhenEveryoneBecomeFriends_1101 solution =
      new TheEarliestMomentWhenEveryoneBecomeFriends_1101();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(
          arguments(
              new int[][] {
                {20190101, 0, 1},
                {20190104, 3, 4},
                {20190107, 2, 3},
                {20190211, 1, 5},
                {20190224, 2, 4},
                {20190301, 0, 3},
                {20190312, 1, 2},
                {20190322, 4, 5}
              },
              6,
              20190301),
          arguments(new int[][] {{0, 2, 0}, {1, 0, 1}, {3, 0, 3}, {4, 1, 2}, {7, 3, 1}}, 4, 3));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void earliestAcq(int[][] logs, int n, int expectedResult) {
    int actualResult = solution.earliestAcq(logs, n);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
